Create User
User Entity - Hopebridge

This repository contains a User entity class for a Spring Boot application. The User entity represents a user in the system and implements Spring Security's UserDetails interface for authentication and authorization purposes. It is built using Jakarta Persistence API (JPA) and utilizes Lombok for reducing boilerplate code.